// ClientAG/src/app/models/scholarship.ts
// export class scholarship {
//     Identity: string;
//     HebrewMonth: string;
//     HebrewYear: string;
//     Date: Date;
//     Basic: number;
//     Musar: boolean;
//     Absence: number;
//     NoteAbsence: string;
//     SumForShmiratSdarim: number;
//     SumForAtmada: number;
//     Atmada: number;
//     NoteAtmada: string;
//     Mark: number;
//     ExcellenceScholarship: number;
//     HolidayScholarship: number;
//     NoteHolidayScholarship: string;
//     ReligiousMinistry: number;
//     Scholarship1: number;
//     ScholarshipsAndLoan: number;
//     Donation: number;
//     Refund: number;
//     RemainderDebt: number;
// }
export class scholarship {
    Identity: string;
    HebrewMonth: string;
    HebrewYear: string;
    Date: Date;
    Basic: number;
    Musar: boolean;
    SumForMusar: number;
    Absence: number;
    NoteAbsence: string;
    SumForAbsence: number;
    ShmiratSdarim: number;
    SumForShmiratSdarim: number;
    addToShmiratSdarimBecauseGoodTest: number;
    Atmada: number;
    NoteAtmada: string;
    SumForAtmada: number;
    Mark: number;
    SumForMark: number;
    ExcellenceScholarship: number;
    HolidayScholarship: number;
    NoteHolidayScholarship: string;
    ReligiousMinistry: number;
    Scholarship: number;
    ScholarshipsAndLoan: number;
    Refund: number;
    Donation: number;
    RemainderDebt: number;
    Confirm: boolean;
    RefundTransportation: number;
    NoteGeneral:string;
}
export class basicScholarship {
    basicFullDay: number;
    musar: number;
    shmiratSdarim: number;
    maxAbsenceForShmiratSdarim: number;
    discountPerAbsenceInMilga: number;
    atmadaFullDay: number;
    fullNumberAtmada: number;
    discountPerAtmada: number;
    test: number;
    minMarkForTestMilga: number;
    addMilgaTestToAtmda: number;
    excellenceScholarship: number;
    minExcellentMarkForMilga: number;
    religiousMinistryForFullDay: number;
    HolidayScholarship: number;
    noteHolidayScholarship: string;
    staticScholarship: number;
    noteGeneral:string;
}